create person w/ role and bugfixes

+++ xray-ejb/src/java/jee18/logic/impl/TimesheetEntrySystem.java
@@ -7,12 +7,15 @@ package jee18.logic.impl;
 
 import java.util.List;
 import javax.ejb.EJB;
+import javax.ejb.EJBException;
 import javax.ejb.Stateless;
 import javax.naming.NamingException;
 import jee18.dao.TimesheetEntryAccess;
 import jee18.dto.Timesheet;
 import jee18.dto.TimesheetEntry;
 import jee18.entities.TimesheetEntryEntity;
+import jee18.entities.enums.ContractStatus;
+import jee18.entities.enums.TimesheetStatus;
 import jee18.logic.AbstractTimesheetSystem;
 import jee18.logic.ITimesheetEntrySystem;
 import jee18.logic.ITimesheetSystem;
@@ -52,11 +55,13 @@ public class TimesheetEntrySystem extends AbstractTimesheetSystem<TimesheetEntry
     @Override
     public TimesheetEntry add(TimesheetEntry te, String timesheetUuid) {
         Timesheet t = timesheetSystem.get(timesheetUuid);
-//        Map<String, AbstractEntity> relations = new HashMap();
-//        relations.put("TIMESHEET", Timesheet.toEntity(t));
-//        return TimesheetEntry.toDTO((TimesheetEntryEntity) timesheetEntryAccess.create(TimesheetEntry.toEntity(te), relations));
-        TimesheetEntryEntity tee = timesheetEntryAccess.createWithTimesheet(TimesheetEntry.toEntity(te), timesheetUuid);
-        return TimesheetEntry.toDTO(tee);
+        if (t.getStatus() == TimesheetStatus.IN_PROGRESS && t.getContract().getStatus() == ContractStatus.STARTED) {
+            TimesheetEntryEntity tee = timesheetEntryAccess.createWithTimesheet(TimesheetEntry.toEntity(te), timesheetUuid);
+            return TimesheetEntry.toDTO(tee);
+        }
+        else {
+            throw new EJBException("Timesheet Entry can only be added when timesheet is in progress and contract was started.");
+        }
     }
 
     @Override
@@ -65,13 +70,24 @@ public class TimesheetEntrySystem extends AbstractTimesheetSystem<TimesheetEntry
     }
 
     @Override
-    public Integer update(String uuid, TimesheetEntry t) {
-        return super.updateByUuid(uuid, t);
+    public Integer update(String uuid, TimesheetEntry te) {
+        if (te.getTimesheet().getStatus() == TimesheetStatus.IN_PROGRESS && te.getTimesheet().getContract().getStatus() == ContractStatus.STARTED) {
+            return super.updateByUuid(uuid, te);
+        }
+        else {
+            throw new EJBException("Timesheet Entry can only be added when timesheet is in progress and contract was started.");
+        }
     }
 
     @Override
     public Integer delete(String uuid) {
-        return super.deleteByUuid(uuid);
+        TimesheetEntry te = super.getByUuid(uuid);
+        if (te.getTimesheet().getStatus() == TimesheetStatus.IN_PROGRESS && te.getTimesheet().getContract().getStatus() == ContractStatus.STARTED) {
+            return super.deleteByUuid(uuid);
+        }
+        else {
+            throw new EJBException("Timesheet Entry can only be added when timesheet is in progress and contract was started.");
+        }
     }
 
 }
